refactor: migrate queries from lib to services directory
- Updated import paths for various query functions to point to the new services directory, improving code organization.
- Removed obsolete query files from the lib directory, streamlining the codebase.
