+++ src/components/landing-page/hero-stats.tsx
@@ -1,6 +1,5 @@
+import { getHeroStats } from "@/services/queries/stats";
 import * as motion from "motion/react-client";
-
-import { getHeroStats } from "@/lib/queries/stats";
 
 import {
   HeaderSectionContainer,
