+++ src/hooks/use-get-user-categories.tsx
@@ -1,5 +1,6 @@
+import { getUserCategories } from "@/services/queries";
+
 import { useServerActionQuery } from "@/hooks/server-action-hooks";
-import { getUserCategories } from "@/lib/queries";
 
 export function useGetUserCategories() {
   return useServerActionQuery(getUserCategories, {
