+++ src/hooks/use-get-user-category-hierarchy.tsx
@@ -1,5 +1,6 @@
+import { getUserCategoryHierarchy } from "@/services/queries";
+
 import { useServerActionQuery } from "@/hooks/server-action-hooks";
-import { getUserCategoryHierarchy } from "@/lib/queries";
 
 export function useGetUserCategoryHierarchy() {
   return useServerActionQuery(getUserCategoryHierarchy, {
