+++ src/hooks/use-get-user-filters.ts
@@ -1,5 +1,6 @@
+import { getFiltersWithItems } from "@/services/queries";
+
 import { useServerActionQuery } from "@/hooks/server-action-hooks";
-import { getFiltersWithItems } from "@/lib/queries";
 
 export function useGetUserFilters() {
   return useServerActionQuery(getFiltersWithItems, {
